requirements.txt
apache-airflow==3.1.7
psycopg2
python-dotenv